Minor test tweaks
