+++ src/scripts/src/reducers/items.test.js
@@ -1,8 +1,7 @@
 import { expect } from 'chai';
-// import * as actions from '../actions/actions';
 import items, { defaultState } from './items';
 
-describe('items reducer', () => {
+describe('reducers/items', () => {
   it('should return same state when no action matches', () => {
     expect(items(defaultState, {})).to.equal(defaultState);
   });
